+++ main/java/Truck.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public abstract class Truck extends Car{
 
     public Attachment attachment;
-    public Truck(int doors, int horsepower, Color carcolor, String model, double x, double y, Attachment attachment) {
+    public Truck(int doors, int horsepower, Color carcolor, String model, int x, int y, Attachment attachment) {
         super(doors, horsepower, carcolor, model, x, y, CarSize.LARGE);
         this.attachment = attachment;
